+++ geotrek/altimetry/management/commands/loaddem.py
@@ -1,3 +1,4 @@
+from django.apps import apps
 from django.contrib.gis.gdal.error import GDALException
 from django.core.management.base import BaseCommand, CommandError
 from django.db import connection
@@ -9,7 +10,6 @@ from subprocess import call, PIPE
 import tempfile
 
 from geotrek.altimetry.models import Dem
-from geotrek.core.models import Path
 
 
 class Command(BaseCommand):
@@ -109,7 +109,10 @@ class Command(BaseCommand):
         if update_altimetry_paths:
             if verbose:
                 self.stdout.write('Updating 3d geometries.\n')
-            Path.objects.all().update(geom=F('geom'))
+            for model in apps.get_models():
+                if 'geom' in [field.name for field in model._meta.get_fields()]:
+                    model.objects.all().update(geom=F('geom'))
+
         return
 
     def call_command_system(self, cmd, **kwargs):